+++ flutter_ffi_plugin/example/native/hub/src/bridge/bridge_engine/mod.rs
@@ -1,13 +1,13 @@
 use std::panic::{RefUnwindSafe, UnwindSafe};
 
 pub use handler::{FfiCallMode, Handler, WrapInfo};
-pub use rust2dart::StreamSink;
+pub use rust_to_dart::StreamSink;
 pub mod ffi;
 pub use ffi::*;
 pub mod handler;
 mod into_into_dart;
 mod macros;
-pub mod rust2dart;
+pub mod rust_to_dart;
 pub mod support;
 
 #[cfg(target_family = "wasm")]
