+++ flutter_ffi_plugin/example/native/hub/src/bridge/bridge_generated.rs
@@ -12,7 +12,7 @@
 // Generated by flutter_rust_bridge_codegen@ 1.80.1.
 
 use crate::bridge::api::*;
-use crate::bridge::bridge_engine::rust2dart::IntoIntoDart;
+use crate::bridge::bridge_engine::rust_to_dart::IntoIntoDart;
 use crate::bridge::bridge_engine::*;
 use core::panic::UnwindSafe;
 use std::ffi::c_void;
@@ -186,7 +186,7 @@ impl support::IntoDart for RustResponse {
     }
 }
 impl support::IntoDartExceptPrimitive for RustResponse {}
-impl rust2dart::IntoIntoDart<RustResponse> for RustResponse {
+impl rust_to_dart::IntoIntoDart<RustResponse> for RustResponse {
     fn into_into_dart(self) -> Self {
         self
     }
@@ -202,7 +202,7 @@ impl support::IntoDart for RustResponseUnique {
     }
 }
 impl support::IntoDartExceptPrimitive for RustResponseUnique {}
-impl rust2dart::IntoIntoDart<RustResponseUnique> for RustResponseUnique {
+impl rust_to_dart::IntoIntoDart<RustResponseUnique> for RustResponseUnique {
     fn into_into_dart(self) -> Self {
         self
     }
@@ -219,7 +219,7 @@ impl support::IntoDart for RustSignal {
     }
 }
 impl support::IntoDartExceptPrimitive for RustSignal {}
-impl rust2dart::IntoIntoDart<RustSignal> for RustSignal {
+impl rust_to_dart::IntoIntoDart<RustSignal> for RustSignal {
     fn into_into_dart(self) -> Self {
         self
     }
